Spawns all reflections on all depths
added recursive method in FindShapePaths to create the shapes of all the lights on all layers. fixed a error where it would try and make a new lightbeam2D with only 2 vertices (these are now ignored). There is still a bug when the next ray has the same number of bounces but from a new mirror