Save and load methods
The Grid manager now holds all the puzzle prefabs and the level button direct to which puzzle from the prefab list. Added a solvedValue int to the puzzle class to track if a puzzle has been solved/ how many guesses it took to solve. The game can now save and load if hard mode is on and an array of all the solvedValues.