Small Changes/Bug fixes
Fixed bug where the continue button wouldn't turn on in the tutorial. Improved the the dragging visual so that the base block layer now also moves the the top when dragging but drops back again when dropped. Swapping difficulty mid level will now reset the level and disables the current difficulty button to avoid accidental resets.